+++ src/treeview/changelog/items/version-tree-item.ts
@@ -3,7 +3,7 @@ import { Changelog } from '../../../classes/changelog';
 import { ChangelogVersion } from '../../../types/changelog';
 
 export class ChangelogVersionTreeItem extends vscode.TreeItem {
-	contextValue?: string | undefined = 'changelog-version';
+	contextValue = 'changelog-version';
 
 	constructor(public readonly changelog: Changelog, public readonly version: ChangelogVersion) {
 		super(version.label, vscode.TreeItemCollapsibleState.Collapsed);