+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - option to open referencing changelog file
 - option to edit changelog version
+- option to add new version to changelog
 
 ### Fixed
 - minor fixes in changelog file export: date will not be exported as `undefined`